Uses new pkg securedrop-workstation-viewer
Supersedes the securedrop-workstation-sd-svs-disp package, to make the
transition to consolidated templates a bit easier, by avoiding dpkg
conflicts altogether.

Tests for securedrop-workstation-svs-disp absence

The package "securedrop-workstation-svs-disp" has been superseded by
"securedrop-workstation-viewer", so in testing for presence of one,
ensure the absence of the other.

Updates the config tests to permit testing package absence; before it
was showing a confusing stack trace if package was missing (so error
rather than failure).