Sets default mimetype handler for sd-proxy
Uses the new sd-mime-handling state and applies it to sd-proxy, so that
symlinks for mimetypes are handled appropriately post-template-
consolidation. Requires packaging changes that are for now only in a
feature branch on the dev packages repo, served via apt-test.

Includes a small fix to ensure that the files in private volume are
owned by normal user, not root.

Removes unused sd-proxy files

The do-not-open here files are left over from a prototype of the
workstation, no longer required.

Tests sd-proxy: use open-in-dvm for mimetypes

We've ditched the do-not-open-here logic, so sd-proxy should now default
to using open-in-dvm for all filetypes. The config tests now reflect
this. Copy/pasted from the sd-app tests, didn't bother to refactor to
make it DRY.