Moves consolidated templates into single state
We originally had the consolidated templates in their own salt state
file, to aid in development, so we could target just that one component.
Folding into the pre-existing state file for the base template now, and
cleaning up the extra references, mostly to minimize the diff and aid in
review.